Add logic, fix typos
I had to somehow bind book objects in array with DOM element in html,
so I used data-id to do it.
In `createDOMBook` I got to create a whole book block. I didn't
want to create by element using `appendChild` bunch of times,
so providing that structure of the book will be the same for every
instance, I just put it in innerHTML.